+++ packages/shared/src/features/opportunity/hooks/useOpportunity.ts
@@ -3,57 +3,12 @@ import type { QueryKey } from '@tanstack/react-query';
 import { RequestKey, StaleTime } from '../../../lib/query';
 import { gqlClient } from '../../../graphql/common';
 import { OPPORTUNITY_BY_ID_QUERY } from '../graphql';
-import type { Organization } from '../../organizations/types';
-import type { PublicProfile } from '../../../lib/user';
-import type { ProtoEnumValue } from '../../../lib/protobuf';
+import type { Opportunity } from '../types';
 
 export const getOpportunityByIdKey = (id: string): QueryKey => [
   RequestKey.Opportunity,
   id,
 ];
-
-type OpportunityContentBlock = {
-  content?: string;
-  html?: string;
-};
-
-export type OpportunityLocation = {
-  type: ProtoEnumValue;
-  city?: string;
-  country?: string;
-  subdivision?: string;
-  continent?: string;
-};
-
-export type Salary = {
-  min: number;
-  max: number;
-  currency: string;
-  period: ProtoEnumValue;
-};
-
-export type Opportunity = {
-  id: string;
-  type: ProtoEnumValue;
-  title: string;
-  tldr: string;
-  organization: Organization;
-  content: {
-    overview: OpportunityContentBlock;
-  };
-  meta: {
-    employmentType: ProtoEnumValue;
-    teamSize: number;
-    salary: Salary;
-    seniorityLevel: ProtoEnumValue;
-    roleType: number;
-  };
-  recruiters: PublicProfile[];
-  location: OpportunityLocation[];
-  keywords?: {
-    value: string;
-  }[];
-};
 
 export const opportunityByIdOptions = ({ id }: { id: string }) => {
   return {
